refactor: 엔티티 중복 제거
