+++ src/main/java/kitchenpos/eatinorders/tobe/domain/EatInOrder.java
@@ -1,29 +1,19 @@
 package kitchenpos.eatinorders.tobe.domain;
 
 import kitchenpos.menus.tobe.domain.Menu;
+import kitchenpos.sharedkernel.Order;
+import kitchenpos.sharedkernel.OrderStatus;
 
-import javax.persistence.*;
+import javax.persistence.Embedded;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Table(name = "eat_in_order")
 @Entity
-public class EatInOrder {
-
-    @Column(name = "id", columnDefinition = "binary(16)")
-    @Id
-    private UUID id;
-
-    @Column(name = "status", nullable = false)
-    @Enumerated(EnumType.STRING)
-    private OrderStatus status;
-
-    @Column(name = "order_date_time", nullable = false)
-    private LocalDateTime orderDateTime;
-
+public class EatInOrder extends Order {
     @Embedded
     private EatInOrderLineItems eatInOrderLineItems;
 
@@ -45,18 +35,6 @@ public class EatInOrder {
         this.orderDateTime = orderDateTime;
         this.eatInOrderLineItems = eatInOrderLineItems;
         this.orderTableId = orderTableId;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public OrderStatus getStatus() {
-        return status;
-    }
-
-    public LocalDateTime getOrderDateTime() {
-        return orderDateTime;
     }
 
     public EatInOrderLineItems getOrderLineItems() {
@@ -90,44 +68,5 @@ public class EatInOrder {
         if (!orderTable.isOccupied()) {
             throw new IllegalStateException();
         }
-    }
-
-    private static Map<UUID, Menu> createMenuMap(final List<Menu> menus) {
-        return menus.stream()
-            .map(menu -> {
-                validateDisplayedMenu(menu);
-                return menu;
-            })
-            .collect(Collectors.toMap(
-                Menu::getId,
-                menu -> menu
-            ));
-    }
-
-    private static void validateDisplayedMenu(final Menu menu) {
-        if (!menu.isDisplayed()) {
-            throw new IllegalStateException();
-        }
-    }
-
-    public void accept() {
-        if (status != OrderStatus.WAITING) {
-            throw new IllegalStateException();
-        }
-        status = OrderStatus.ACCEPTED;
-    }
-
-    public void serve() {
-        if (status != OrderStatus.ACCEPTED) {
-            throw new IllegalStateException();
-        }
-        status = OrderStatus.SERVED;
-    }
-
-    public void complete() {
-        if (status != OrderStatus.SERVED) {
-            throw new IllegalStateException();
-        }
-        status = OrderStatus.COMPLETED;
     }
 }
