+++ src/main/java/kitchenpos/eatinorders/tobe/domain/EatInOrderRepository.java
@@ -1,5 +1,7 @@
 package kitchenpos.eatinorders.tobe.domain;
 
+import kitchenpos.sharedkernel.OrderStatus;
+
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
