--- src/main/java/kitchenpos/eatinorders/tobe/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, COMPLETED
-}
