+++ src/main/java/kitchenpos/sharedkernel/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.deliveryorders.tobe.domain;
+package kitchenpos.sharedkernel;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
