--- src/main/java/kitchenpos/takeoutorders/tobe/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.takeoutorders.tobe.domain;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, COMPLETED
-}
