+++ src/main/java/kitchenpos/takeoutorders/tobe/domain/TakeOutOrder.java
@@ -1,28 +1,19 @@
 package kitchenpos.takeoutorders.tobe.domain;
 
 import kitchenpos.menus.tobe.domain.Menu;
+import kitchenpos.sharedkernel.Order;
+import kitchenpos.sharedkernel.OrderStatus;
 
-import javax.persistence.*;
+import javax.persistence.Embedded;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Table(name = "take_out_order")
 @Entity
-public class TakeOutOrder {
-
-    @Column(name = "id", columnDefinition = "binary(16)")
-    @Id
-    private UUID id;
-
-    @Column(name = "status", nullable = false)
-    @Enumerated(EnumType.STRING)
-    private OrderStatus status;
-
-    @Column(name = "order_date_time", nullable = false)
-    private LocalDateTime orderDateTime;
+public class TakeOutOrder extends Order {
 
     @Embedded
     private OrderLineItems orderLineItems;
@@ -43,22 +34,6 @@ public class TakeOutOrder {
         this.orderLineItems = orderLineItems;
     }
 
-    public UUID getId() {
-        return id;
-    }
-
-    public LocalDateTime getOrderDateTime() {
-        return orderDateTime;
-    }
-
-    public OrderLineItems getOrderLineItems() {
-        return orderLineItems;
-    }
-
-    public OrderStatus getStatus() {
-        return status;
-    }
-
     public static TakeOutOrder create(
         final List<Menu> menus,
         final List<TakeOutOrderLineItem> takeOutOrderLineItemList
@@ -69,44 +44,5 @@ public class TakeOutOrder {
             LocalDateTime.now(),
             new OrderLineItems(takeOutOrderLineItemList, createMenuMap(menus))
         );
-    }
-
-    private static Map<UUID, Menu> createMenuMap(final List<Menu> menus) {
-        return menus.stream()
-            .map(menu -> {
-                validateDisplayedMenu(menu);
-                return menu;
-            })
-            .collect(Collectors.toMap(
-                Menu::getId,
-                menu -> menu
-            ));
-    }
-
-    private static void validateDisplayedMenu(final Menu menu) {
-        if (!menu.isDisplayed()) {
-            throw new IllegalStateException();
-        }
-    }
-
-    public void accept() {
-        if (status != OrderStatus.WAITING) {
-            throw new IllegalStateException();
-        }
-        status = OrderStatus.ACCEPTED;
-    }
-
-    public void serve() {
-        if (status != OrderStatus.ACCEPTED) {
-            throw new IllegalStateException();
-        }
-        status = OrderStatus.SERVED;
-    }
-
-    public void complete() {
-        if (status != OrderStatus.SERVED) {
-            throw new IllegalStateException();
-        }
-        status = OrderStatus.COMPLETED;
     }
 }
