+++ src/test/java/kitchenpos/deliveryorders/tobe/application/DeliveryOrderServiceTest.java
@@ -5,12 +5,12 @@ import kitchenpos.deliveryorders.tobe.application.dto.OrderLineItemRequest;
 import kitchenpos.deliveryorders.tobe.domain.DeliveryAddress;
 import kitchenpos.deliveryorders.tobe.domain.DeliveryOrder;
 import kitchenpos.deliveryorders.tobe.domain.DeliveryOrderRepository;
-import kitchenpos.deliveryorders.tobe.domain.OrderStatus;
 import kitchenpos.deliveryorders.tobe.infra.FakeKitchenridersClient;
 import kitchenpos.deliveryorders.tobe.infra.InMemoryDeliveryOrderRepository;
 import kitchenpos.menus.application.InMemoryMenuRepository;
 import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.sharedkernel.OrderStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
