+++ src/test/java/kitchenpos/deliveryorders/tobe/domain/DeliveryOrderTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.deliveryorders.tobe.domain;
 
+import kitchenpos.sharedkernel.OrderStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
