+++ src/test/java/kitchenpos/takeoutorders/tobe/application/TakeOutOrderServiceTest.java
@@ -3,9 +3,9 @@ package kitchenpos.takeoutorders.tobe.application;
 import kitchenpos.menus.application.InMemoryMenuRepository;
 import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.sharedkernel.OrderStatus;
 import kitchenpos.takeoutorders.tobe.application.dto.OrderLineItemRequest;
 import kitchenpos.takeoutorders.tobe.application.dto.TakeOutOrderRequest;
-import kitchenpos.takeoutorders.tobe.domain.OrderStatus;
 import kitchenpos.takeoutorders.tobe.domain.TakeOutOrder;
 import kitchenpos.takeoutorders.tobe.domain.TakeOutOrderRepository;
 import kitchenpos.takeoutorders.tobe.infra.InMemoryTakeOutOrderRepository;
